backend/pyproject.toml
[project]
name = "ordertogether-backend"
version = "0.1.0"
description = "OrderTogether API"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.5",
    "uvicorn[standard]>=0.32.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.12",
]

[tool.uv]
dev-dependencies = []
